Add a Levenshtein distance implementation
The Levenshtein distance between two strings is the amount of additions,
substractions and substitutions required to transform one string in to
another. It's useful to determine if two strings are similar, such as
when users append dates to passwords when corporate policy requires
regular password changes, or when the perform 'l33t' transformations to
turn their 'bad-passwords' into 'b4d-p4ssw0rd$' that pass websites
password requirements, but are not really any more secure.

Issue: #551